+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New `re create annotations` command for uploading annotations (labels and
   entities) to existing comments in a dataset, without having to use `re create comments`.
   This avoids potentially - and unknowingly - modifying the underlying comments in the source.
+- Add support to `--force` delete projects with existing resources.
 
 # v.0.9.0